+++ frontend/apps/marketing/src/components/common/utils/index.ts
